A lot of changes in one commit, which is bad form. :(
To summarize:

* Reformat and restructure a few things. This is now sorta-kinda-almost actually
a playbook, which means the hosts and tasks heading should (and does) go away,
and all tasks aren't nested + indented by default.
* Factored out the username (merlin) out to a variable that's set in
defaults/main.yml for both devenv and desktop.
* Also pulled out the js var to the same file.
* Added some tasks to configure the repo for Brave and install it on the desktop
side of things.
* Created a playbook.yml file at the root that calls both roles and targets them
to servers in either development or desktop roles. At least on my fedora WSL,
this does the right thing and runs the devenv, but skips the desktop one. I'll
have to test on my fedora bare metal install and see if I can configure both to
target localhost.

Still to do:

* Decompose things down to single-task. Set up dotfiles, install emacs, get
spacemacs, that kinda thing. Very little of this is as straightforward as
"install nginx and copy this config over." There are many dependencies to sort
out.
* Handle other distros?
* Publish to ansible-galaxy?
* Leverage existing tasks/roles on same?